run/unzip/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
py7zr==0.20.0
google-cloud-storage==2.5.0
